# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateTaskRequest:
    properties:
      description:
        maxLength: 5000
        type: string
      due_date:
        type: string
      priority:
        enum:
        - low
        - medium
        - high
        - urgent
        type: string
      status_id:
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - priority
    - status_id
    - title
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.ReorderTasksRequest:
    properties:
      after_id:
        description: ID da tarefa após a qual inserir (null = início)
        type: string
      status_id:
        type: string
      task_id:
        type: string
    required:
    - status_id
    - task_id
    type: object
...
  dto.TaskDTO:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      lexorank:
        type: string
      priority:
        type: string
      status_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  dto.UpdateTaskRequest:
    properties:
      description:
        maxLength: 5000
        type: string
      due_date:
        type: string
      priority:
        enum:
        - low
        - medium
        - high
        - urgent
        type: string
      status_id:
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - priority
    - status_id
    - title
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API do sistema Organiza Aqui
  title: Organiza Aqui API
  version: "1.0"
paths:
...
  /api/v1/tasks:
    get:
      description: Retorna todas as tarefas do usuário com filtros opcionais
      parameters:
      - description: Filtrar por status
        in: query
        name: status_id
        type: string
      - description: Filtrar por prioridade
        in: query
        name: priority
        type: string
      - description: Filtrar por completado
        in: query
        name: completed
        type: boolean
      - description: Limite de resultados
        in: query
        name: limit
        type: integer
      - description: Offset para paginação
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TaskDTO'
            type: array
      security:
      - BearerAuth: []
      summary: Listar tarefas
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Cria uma nova tarefa
      parameters:
      - description: Dados da tarefa
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TaskDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar tarefa
      tags:
      - tasks
  /api/v1/tasks/{id}:
    delete:
      description: Remove uma tarefa
      parameters:
      - description: ID da tarefa
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deletar tarefa
      tags:
      - tasks
    get:
      description: Retorna uma tarefa específica
      parameters:
      - description: ID da tarefa
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obter tarefa
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Atualiza os dados de uma tarefa
      parameters:
      - description: ID da tarefa
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar tarefa
      tags:
      - tasks
  /api/v1/tasks/{id}/complete:
    post:
      description: Marca uma tarefa como completa
      parameters:
      - description: ID da tarefa
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskDTO'
      security:
      - BearerAuth: []
      summary: Completar tarefa
      tags:
      - tasks
  /api/v1/tasks/{id}/uncomplete:
    post:
      description: Marca uma tarefa como incompleta
      parameters:
      - description: ID da tarefa
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskDTO'
      security:
      - BearerAuth: []
      summary: Descompletar tarefa
      tags:
      - tasks
  /api/v1/tasks/reorder:
    post:
      consumes:
      - application/json
      description: Reordena uma tarefa usando drag-and-drop
      parameters:
      - description: Dados de reordenação
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReorderTasksRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskDTO'
      security:
      - BearerAuth: []
      summary: Reordenar tarefa
      tags:
      - tasks
...
swagger: "2.0"